# springboot/build/resources/main/openapi/test-api.yaml
openapi: 3.0.1
info:
  title: PPAP Chat Service API
  version: 1.0.0
servers:
  - url: http://localhost:8080 #기초 경로 springboot
...
paths:
...
  /api/users/signup:
    post:
      tags:
        - users
      summary: 사용자 가입
      description: 새로운 사용자를 가입합니다.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: 가입 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: 가입 실패
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Invalid user data provided
  /api/users/login:
    post:
      tags:
        - users
      summary: 사용자 로그인
      description: 사용자 로그인 처리
      operationId: userLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserToken'
        '400':
          description: 로그인 실패
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Invalid user data provided
...
components:
  schemas:
...
    User:
      title: User Aggregate
      description: user info
      type: object
      required: 
        - loginId
        - username
        - password
        - confirmPassword
        - email
        - userType
      properties:
        userId:
          type: integer
          format: int64
          description: user 테이블 pk
        loginId:
          type: string
          description: 사용자 아이디
        username:
          type: string
          description: 사용자 이름
        password:
          type: string
          description: 사용자 비밀번호
        confirmPassword: 
          type: string
          description: 사용자 비밀번호 확인
        email:
          type: string
          description: 사용자 이메일
        isActive:
          type: boolean
          description: 사용 여부
          default: true
        isWithdrawed:
          type: boolean
          description: 탈퇴 여부
          default: false
        userType:
          type: string
          enum:
            - INDI
            - CORP

        audit:
          $ref: '#/components/schemas/Audit'
    UserToken:
      type: object
      description: 사용자 json web token
      required:
        - accessToken
        - refreshToken
      properties:
        accessToken:
          type: string
          description: access token
        refreshToken:
          type: string
          description: refresh token
    Audit:
      type: object
      properties:
        createdDate:
          type: string
          format: date-time
          description: 생성일
        createdBy:
          type: string
          description: 생성자
        modifiedDate:
          type: string
          format: date-time
          description: 수정일
        modifiedBy:
          type: string
          description: 수정자
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []